# packages/api-sdk/src/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: string;
};

export type UserBody = {
    name: string;
    email: string;
    password: string;
};

export type UserView = {
    id: number;
    name: string;
    email: string;
};

export type LoginBody = {
    email: string;
    password: string;
};

export type ListUsersData = {
    body?: never;
    path?: never;
    query?: {
        /**
         * Filter users by name or email
         */
        search?: string;
    };
    url: '/api/users';
};

export type ListUsersResponses = {
    /**
     * Users retrieved successfully
     */
    200: Array<UserView>;
};

export type ListUsersResponse = ListUsersResponses[keyof ListUsersResponses];

export type CreateUserData = {
    body: UserBody;
    path?: never;
    query?: never;
    url: '/api/users';
};

export type CreateUserResponses = {
    /**
     * User created successfully
     */
    201: UserView;
};

export type CreateUserResponse = CreateUserResponses[keyof CreateUserResponses];

export type DeleteUserByIdData = {
    body?: never;
    path: {
        id: string;
    };
    query?: never;
    url: '/api/users/{id}';
};

export type DeleteUserByIdResponses = {
    /**
     * User deleted successfully
     */
    204: void;
};

export type DeleteUserByIdResponse = DeleteUserByIdResponses[keyof DeleteUserByIdResponses];

export type GetUserByIdData = {
    body?: never;
    path: {
        id: string;
    };
    query?: never;
    url: '/api/users/{id}';
};

export type GetUserByIdResponses = {
    /**
     * User retrieved successfully
     */
    200: UserView;
};

export type GetUserByIdResponse = GetUserByIdResponses[keyof GetUserByIdResponses];

export type UpdateUserByIdData = {
    body: UserBody;
    path: {
        id: string;
    };
    query?: never;
    url: '/api/users/{id}';
};

export type UpdateUserByIdResponses = {
    /**
     * User updated successfully
     */
    200: UserView;
};

export type UpdateUserByIdResponse = UpdateUserByIdResponses[keyof UpdateUserByIdResponses];

export type LoginData = {
    body: LoginBody;
    path?: never;
    query?: never;
    url: '/api/auth/login';
};

export type LoginResponses = {
    200: unknown;
};